Moved all functions out of main. Removed duplicate dependencies from modules. Fixed up minor bugs Removed most compile warnings.
Big thanks to [this](http://rustbyexample.com/mod/split.html).
